Sentinel and sanity check if step context is accidentally dropped
This adds a sentinel to the world/context that is threaded between
steps. The sentinel is used to check whether a step implementation
failed to return an updated context. This is likely a mistake if it
happens, and will usually lead to errors farther along in a test that
will be hard to troubleshoot.